ai-assistant/frontend/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-SocketIO==5.3.5
Flask-Cors==4.0.0
Flask-Session==0.5.0

# Authentication
authlib==1.2.1
requests==2.31.0

# HTTP clients
httpx==0.25.2

# WebSocket
python-socketio==5.10.0
eventlet==0.33.3

# Templates and forms
Jinja2==3.1.2
WTForms==3.1.1
Flask-WTF==1.2.1

# Session management
redis==5.0.1

# Environment management
python-dotenv==1.0.0

# Security
cryptography==41.0.7
PyJWT==2.8.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
black==23.12.0
flake8==6.1.0
pytest==7.4.3
pytest-flask==1.3.0